Use PartitionEvaluator in DataSourceScanExec
